+++ src/aggregation/intermediate_agg_result.rs
@@ -36,7 +36,10 @@ pub struct IntermediateAggregationResults {
 
 impl IntermediateAggregationResults {
     /// Convert intermediate result and its aggregation request to the final result.
-    pub fn into_final_bucket_result(self, req: Aggregations) -> crate::Result<AggregationResults> {
+    pub(crate) fn into_final_bucket_result(
+        self,
+        req: Aggregations,
+    ) -> crate::Result<AggregationResults> {
         self.into_final_bucket_result_internal(&(req.into()))
     }