+++ src/schema/mod.rs
@@ -117,6 +117,7 @@ mod field_type;
 mod field_value;
 
 mod bytes_options;
+mod date_time_options;
 mod field;
 mod flags;
 mod index_record_option;
@@ -127,6 +128,7 @@ mod text_options;
 mod value;
 
 pub use self::bytes_options::BytesOptions;
+pub use self::date_time_options::{DateOptions, DatePrecision};
 pub use self::document::Document;
 pub(crate) use self::facet::FACET_SEP_BYTE;
 pub use self::facet::{Facet, FacetParseError};